Added original sqlclient zips, refactores install code and removed echo statements.

+++ src/plugins/ansible-openlink/ansible-openlink.sh
@@ -42,7 +42,7 @@ ansible-openlink_main()
             mkdir -p ${workspace}/${repo}/library
         fi
 
-        if [[ ${repo} == "ansible-module-devel" ]]
+        if [[ ${repo} == "ansible-module-devel" ]] && [ ! -L  ${library}/${repo} ]
         then
             ln -s ${workspace}/${repo}/src ${library}/${repo}
         fi
